buildkite artifact search: renames to satisfy linter
